Move insert observation splitting to splitter/merger
